+++ src/clipper_wrap.cpp
@@ -79,7 +79,7 @@ export void cl_offset_free(ClipperOffset *co) {
 	delete co;
 }
 
-export Paths* cl_offset_path(ClipperOffset* co,Path *subj,int offset,int joinType,int endType) {
+export Paths* cl_offset_path(ClipperOffset* co,Path *subj,double offset,int joinType,int endType) {
   Paths *solution = new Paths();
   co->AddPath(*subj,JoinType(joinType),EndType(endType));
 	// XXX check execute return value?
@@ -87,7 +87,7 @@ export Paths* cl_offset_path(ClipperOffset* co,Path *subj,int offset,int joinTyp
   return solution;
 }
 
-export Paths* cl_offset_paths(ClipperOffset* co,Paths *subj,int offset,int joinType,int endType) {
+export Paths* cl_offset_paths(ClipperOffset* co,Paths *subj,double offset,int joinType,int endType) {
   Paths *solution = new Paths();
   co->AddPaths(*subj,JoinType(joinType),EndType(endType));
   co->Execute(*solution,offset);
